backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=20.1.0
whitenoise>=6.5.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
python-dotenv>=1.0.0
redis>=4.5.0
celery>=5.3.0
channels>=4.0.0
channels-redis>=4.0.0
djangorestframework-simplejwt>=5.3.0
pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
pywhatkit>=5.4
twilio>=8.2.0